+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v0.6.0 (in development)
 -----------------------
+- Added ANSI color escape sequence support
 
 v0.5.0 (2018-07-06)
 -------------------